Struct cranelift_filetests::SingleFunctionCompiler

pub struct SingleFunctionCompiler { /* fields omitted */ }

Compile a single function.

Several Cranelift functions need the ability to run Cranelift IR (e.g. test_run); this SingleFunctionCompiler provides a way for compiling Cranelift Functions to [CompiledFunction]s and subsequently calling them through the use of a [Trampoline]. As its name indicates, this compiler is limited: any functionality that requires knowledge of things outside the Function will likely not work (e.g. global values, calls). For an example of this "outside-of-function" functionality, see cranelift_simplejit::backend::SimpleJITBackend.

use cranelift_filetests::SingleFunctionCompiler;
use cranelift_reader::parse_functions;

let code = "test run \n function %add(i32, i32) -> i32 {  block0(v0:i32, v1:i32):  v2 = iadd v0, v1  return v2 }".into();
let func = parse_functions(code).unwrap().into_iter().nth(0).unwrap();
let mut compiler = SingleFunctionCompiler::with_default_host_isa();
let compiled_func = compiler.compile(func).unwrap();
println!("Address of compiled function: {:p}", compiled_func.as_ptr());

Implementations

impl SingleFunctionCompiler

pub fn new(isa: Box<dyn TargetIsa>) -> Self

Build a SingleFunctionCompiler from a TargetIsa. For functions to be runnable on the host machine, this [TargetISA] must match the host machine's ISA (see [with_host_isa]).

pub fn with_host_isa(flags: Flags) -> Self

Build a SingleFunctionCompiler using the host machine's ISA and the passed flags.

pub fn with_default_host_isa() -> Self

Build a SingleFunctionCompiler using the host machine's ISA and the default flags for this ISA.

pub fn compile(
&mut self,
function: Function
) -> Result<CompiledFunction, CompilationError>

Compile the passed Function to a [CompiledFunction]. This function will:

check that the default ISA calling convention is used (to ensure it can be called)

compile the Function

compile a [Trampoline] for the Function's signature (or used a cached [Trampoline]; this makes it possible to call functions when the signature is not known until runtime.