Struct cranelift_filetests::function_runner::TestFileCompiler

pub struct TestFileCompiler { /* private fields */ }

Compile a test case.

Several Cranelift functions need the ability to run Cranelift IR (e.g. test_run); this TestFileCompiler provides a way for compiling Cranelift Functions to CompiledFunctions and subsequently calling them through the use of a Trampoline. As its name indicates, this compiler is limited: any functionality that requires knowledge of things outside the Function will likely not work (e.g. global values, calls). For an example of this “outside-of-function” functionality, see cranelift_jit::backend::JITBackend.

use cranelift_filetests::TestFileCompiler;
use cranelift_reader::parse_functions;
use cranelift_codegen::data_value::DataValue;

let code = "test run \n function %add(i32, i32) -> i32 {  block0(v0:i32, v1:i32):  v2 = iadd v0, v1  return v2 }".into();
let func = parse_functions(code).unwrap().into_iter().nth(0).unwrap();
let mut compiler = TestFileCompiler::with_default_host_isa().unwrap();
compiler.declare_function(&func).unwrap();
compiler.define_function(func.clone(), ctrl_plane).unwrap();
compiler.create_trampoline_for_function(&func, ctrl_plane).unwrap();
let compiled = compiler.compile().unwrap();
let trampoline = compiled.get_trampoline(&func).unwrap();

let returned = trampoline.call(&vec![DataValue::I32(2), DataValue::I32(40)]);
assert_eq!(vec![DataValue::I32(42)], returned);

Implementations

impl TestFileCompiler

pub fn new(isa: OwnedTargetIsa) -> Self

Build a TestFileCompiler from a TargetIsa. For functions to be runnable on the host machine, this TargetIsa must match the host machine’s ISA (see TestFileCompiler::with_host_isa).

pub fn with_host_isa(flags: Flags) -> Result<Self>

Build a TestFileCompiler using the host machine’s ISA and the passed flags.

pub fn with_default_host_isa() -> Result<Self>

Build a TestFileCompiler using the host machine’s ISA and the default flags for this ISA.

pub fn add_functions( &mut self, functions: &[Function], ctrl_planes: Vec<ControlPlane> ) -> Result<()>

Declares and compiles all functions in functions. Additionally creates a trampoline for each one of them.

pub fn add_testfile(&mut self, testfile: &TestFile<'_>) -> Result<()>

Registers all functions in a [TestFile]. Additionally creates a trampoline for each one of them.

pub fn declare_function(&mut self, func: &Function) -> Result<()>

Declares a function an registers it as a linkable and callable target internally

pub fn define_function( &mut self, func: Function, ctrl_plane: &mut ControlPlane ) -> Result<()>

Defines the body of a function

pub fn create_trampoline_for_function( &mut self, func: &Function, ctrl_plane: &mut ControlPlane ) -> Result<()>

Creates and registers a trampoline for a function if none exists.

pub fn compile(self) -> Result<CompiledTestFile, CompilationError>

Finalize this TestFile and link all functions.